+++ http/index.ts
@@ -1,11 +1,11 @@
-import { makeTaskDriver, TaskSource, TaskDriver } from '@cycler/task'
+import { makeTaskDriver, InputTaskSource, TaskDriver } from '@cycler/task'
 import { Stream } from 'xstream'
 import { getProgressiveResponse, normalizeRequest } from './getResponse'
 import {
   HTTPRequest, HTTPRequestInput, HTTPResponse
 } from './interfaces'
 
-export type HTTPSource = TaskSource<HTTPRequest, HTTPResponse>
+export type HTTPSource = InputTaskSource<HTTPRequestInput, HTTPRequest, HTTPResponse>
 
 export function makeHTTPDriver() {
   return makeTaskDriver