+++ task/helpers/common.ts
@@ -1,6 +1,9 @@
+import { requestOps } from '../requestOps'
 
 export const setCategory = <C extends string>(category: C) =>
-  <T>(obj: T): T & { category: C } => Object.assign({}, obj, { category })
+  <T>(obj: T): T & { category: C } => requestOps
+    .addProperty(obj, 'category', category) as any
 
 export const makeLazy = <T>(obj: T): T & { lazy: boolean } =>
-  Object.assign({}, obj, { lazy: true })
+  requestOps.addProperty(obj, 'lazy', true) as any
+